--- gatsby-test/gatsby-config.js
@@ -1,18 +1,0 @@
-/** @type {import('gatsby').GatsbyConfig} */
-module.exports = {
-  siteMetadata: {
-    title: ``,
-    siteUrl: `https://www.yourdomain.tld`,
-  },
-  plugins: [
-    "gatsby-transformer-remark",
-    {
-      resolve: "gatsby-source-filesystem",
-      options: {
-        name: "pages",
-        path: "../content/",
-      },
-      __key: "pages",
-    },
-  ],
-};
