--- gatsby-test/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "contentlayer-benchmark-gatsby-test",
-  "version": "0.0.1",
-  "private": true,
-  "description": "gatsby",
-  "author": "Sean C Davis",
-  "keywords": [
-    "gatsby"
-  ],
-  "scripts": {
-    "develop": "gatsby develop",
-    "start": "gatsby develop",
-    "build": "gatsby build",
-    "serve": "gatsby serve",
-    "clean": "gatsby clean",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "gatsby": "^4.9.2",
-    "gatsby-source-filesystem": "^4.9.0",
-    "gatsby-transformer-remark": "^5.9.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
-  },
-  "devDependencies": {
-    "@types/node": "^17.0.21",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
-    "typescript": "^4.6.2"
-  }
-}
