--- gatsby-test/src/types/Page.ts
@@ -1,11 +1,0 @@
-export type PageFrontmatter = {
-  title: string;
-};
-
-export type Page = {
-  frontmatter: PageFrontmatter;
-  html: string;
-  fields: {
-    url: string;
-  };
-};
