+++ gatsby/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "contentlayer-benchmark-gatsby",
+  "name": "contentlayer-benchmark-gatsby-test",
   "version": "0.0.1",
   "private": true,
   "description": "gatsby",
