+++ gatsby/src/types/Page.ts
@@ -5,4 +5,7 @@ export type PageFrontmatter = {
 export type Page = {
   frontmatter: PageFrontmatter;
   html: string;
+  fields: {
+    url: string;
+  };
 };
