+++ package.json
@@ -3,7 +3,6 @@
   "name": "contentlayer-benchmark-gatsby-docs",
   "workspaces": [
     "gatsby",
-    "gatsby-test",
     "next-contentlayer",
     "next-remark",
     "scripts/*"
